Switch to the Apache 2 license
See this thread on GitHub for details:
https://github.com/dhale/jtk/issues/10#issuecomment-143578005

Steps followed:

1) The old copyright headers were stripped using:

       perl -0777 -i -pe 's/\/\*\*\**\n\s*Copyright \(c\) \d\d\d\d, Colorado School of Mines.*cpl-v10.html\n\*\*\**\/?\n//igs' $(git grep -l 'Copyright')

2) New headers were added using:

       gradle licenseFormatMain

   Subsequent runs of this gradle command will update the headers,
   keeping the copyrights current in future years.

3) The package.html files were manually tweaked to put the DOCTYPE
   back on the first line, before the copyright header.

4) The license.txt file was updated to APLv2.

The copyright date range "2003 - ${year}" is now used because:

* 2003 is the earliest copyright year I saw within Mines JTK.

* Use of a year range, rather than single year, creates consistency
  across the project which makes it easier to update in the future,
  as well as easier to check for license header inconsistencies.

That said, advice online is inconsistent in its recommendations
regarding whether to use a copyright year range, or date of first
publication, within the copyright header. This commit pragmatically
opts for a year range so the license-gradle-plugin tooling can be
used to quickly maintain the license headers now and in the future.